Move discoverFSCreated/DeletedProjects in the projects module
