+++ services/orchest-webserver/app/app/core/projects.py
@@ -7,7 +7,7 @@ from typing import Optional
 import requests
 from flask.globals import current_app
 
-from _orchest.internals.two_phase_executor import TwoPhaseFunction
+from _orchest.internals.two_phase_executor import TwoPhaseExecutor, TwoPhaseFunction
 from app import error
 from app.connections import db
 from app.core.pipelines import AddPipelineFromFS, DeletePipeline
@@ -403,3 +403,70 @@ def build_environments_for_project(project_uuid):
     return build_environments(
         [environment.uuid for environment in environments], project_uuid
     )
+
+
+def discoverFSDeletedProjects():
+    """Cleanup projects that were deleted from the filesystem."""
+
+    project_paths = [
+        entry.name
+        for entry in os.scandir(current_app.config["PROJECTS_DIR"])
+        if entry.is_dir()
+    ]
+
+    fs_removed_projects = Project.query.filter(
+        Project.path.notin_(project_paths),
+        # This way we do not delete a project that is already being
+        # deleted twice, and avoid considering a project that is
+        # being initialized as deleted from the filesystem, or that
+        # it is moving.
+        Project.status.in_(["READY"]),
+    ).all()
+
+    # Use a TwoPhaseExecutor for each project so that issues in one
+    # project do not hinder the deletion of others.
+    for proj_uuid in [project.uuid for project in fs_removed_projects]:
+        try:
+            with TwoPhaseExecutor(db.session) as tpe:
+                DeleteProject(tpe).transaction(proj_uuid)
+        except Exception as e:
+            current_app.logger.error(
+                ("Error during project deletion (discovery) of " f"{proj_uuid}: {e}.")
+            )
+
+
+def discoverFSCreatedProjects():
+    """Detect projects that were added through the file system."""
+
+    # Detect new projects by detecting directories that were not
+    # registered in the db as projects.
+    existing_project_paths = [project.path for project in Project.query.all()]
+    project_paths = [
+        entry.name
+        for entry in os.scandir(current_app.config["PROJECTS_DIR"])
+        if entry.is_dir()
+    ]
+    new_project_paths = set(project_paths) - set(existing_project_paths)
+
+    # Do not do project discovery if a project is moving, because
+    # a new_project_path could actually be a moving project. Given
+    # that a project has no (persisted) uuid, we won't be able to
+    # find if a new_project_path is actually a MOVING project. NOTE:
+    # we could wait in this endpoint until there is no more MOVING
+    # project.
+    if Project.query.filter(Project.status.in_(["MOVING"])).count() > 0:
+        return
+
+    # Use a TwoPhaseExecutor for each project so that issues in one
+    # project do not hinder the discovery of others.
+    for new_project_path in new_project_paths:
+        try:
+            with TwoPhaseExecutor(db.session) as tpe:
+                CreateProject(tpe).transaction(new_project_path)
+        except Exception as e:
+            current_app.logger.error(
+                (
+                    "Error during project initialization (discovery) of "
+                    f"{new_project_path}: {e}."
+                )
+            )
